Ensure that error issues have integer location values
The Code Climate Platform expects that engines emit location data with integer
values only. Radon is in complete compliance with the spec with the slight
exception of the error result that generates a custom issue from the error
details.

This commit casts the captured line number (grabbed from the error content) to
an int, before it's passed off to the issue formatter.