Hackish fix for #127.
There is probably a better fix long-term, but this makes `mi` work again
for 90%+ of our code base (which have an encoding statement at the top).

The fix grabs the filename from the (allready opened) `fobj` and reads
its contents as binary, leaving it up to `ast.parse(..)` to deal with
the encoding statement.