fix(code-climate): allow python version detection from strings in config (#134)
This was only considering config as string vals, which was fine on
platform where `CC::Yaml` mutated all inner YAML values to strings. That
was, in general, undesirable behavior we didn't want, but we were
relying on it here.

This now accepts either string or numerical values from the config.