Make the build reproducible.
Whilst working on the Reproducible Builds effort [0], we noticed
that radon could not be built reproducibly. This is due to the
documentation including the current build date.

This was originally filed in Debian as #912152 [1].

The attached patch uses the SOURCE_DATE_EPOCH environment variable,
more info of which can be found at [2].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/912152
 [2] https://reproducible-builds.org/specs/source-date-epoch/

Signed-off-by: Chris Lamb <lamby@debian.org>